requirements.txt
matplotlib==3.7.1
nltk==3.8.1
numpy==1.23.4
pandas==1.5.3
Pillow>=9.4.0
scikit_learn==1.2.2
spacy==3.6.1
streamlit==1.28.1
tensorflow==2.12.*
tqdm==4.65.0
